perf: don't load Query Import on home page
